Added staking support
Kraken's public REST API supports 5 stacking endpoints:
- get stackable assets
- stake
- unstake
- get recent transactions
- get pending transactions
The first one should be used to retrieve the necessary
parameters that need to be passed when staking or
unstaking.
The last 2 APIs let you monitor the (un)staking requests,
though the kraken API is a bit unintuitive - the refid of
those requests are not present in the transactions API.
The refid is a ledger id.

This commit also adds a new LedgerEntryType - Staking.